gcc warning fixes

+++ external/chromium_zlib/fast_crc32.cpp
@@ -82,7 +82,7 @@ void check_cpuid(int function_id, int32_t (&data)[4]) {
 #else
     uint32_t max_level, t2, t3, t4;
     __get_cpuid(0, &max_level, &t2, &t3, &t4);
-    if (max_level < function_id) { data[0] = 0; data[1] = 0; data[2] = 0; data[3] = 0; return; }
+    if (max_level < static_cast<uint32_t>(function_id)) { data[0] = 0; data[1] = 0; data[2] = 0; data[3] = 0; return; }
     __get_cpuid(function_id,
             reinterpret_cast<uint32_t*>(data[0]),
             reinterpret_cast<uint32_t*>(data[1]),
